:wrench: Refactor

+++ src/index.ts
@@ -1,7 +1,3 @@
-if (!fs.existsSync(path.join(__dirname, "../spotify.json"))) {
-	console.log("Please run npm run-script authorize");
-	process.exit();
-}
 import axios from "axios";
 import fastify from "fastify";
 import fs from "fs";
@@ -9,45 +5,47 @@ import path from "path";
 
 import config from "../config.json";
 import Lookup from "./inc/lookup";
-import Lyrics from "./inc/lyrics";
 
-const started = new Date();
+if (!fs.existsSync(path.join(__dirname, "../spotify.json"))) {
+	console.log("Please run npm run-script authorize");
+	process.exit();
+}
 
-const lookup = new Lookup({
-	clientId: config.spotify_client_id,
-	clientSecret: config.spotify_client_secret,
-	redirectUri: config.spotify_redirect
-});
-const lyrics = new Lyrics();
-
-const server = fastify();
+const started: any = new Date(),
+	lookup = new Lookup({
+		clientId: config.spotify_client_id,
+		clientSecret: config.spotify_client_secret,
+		redirectUri: config.spotify_redirect
+	}),
+	server = fastify();
 
 server.addHook('onResponse', (req, res, done) => {
-    let stats = JSON.parse(<any>fs.readFileSync(path.join(__dirname, "../stats.json")));
-    stats.requests++;
-    fs.writeFileSync(path.join(__dirname, "../stats.json"), JSON.stringify(stats));
-    done()
+	let stats = JSON.parse(<any>fs.readFileSync(path.join(__dirname, "../stats.json")));
+	stats.requests++;
+	fs.writeFileSync(path.join(__dirname, "../stats.json"), JSON.stringify(stats));
+	done()
 })
 
 server.get('/api', (req, res) => {
-	let endTime = <any>new Date();
-  	var timeDiff = endTime - config.started; //in ms
-  	// strip the ms
-  	timeDiff /= 1000;
+	let endTime: any = new Date(),
+		timeDiff = endTime - started;
 
-	var sec_num = Math.round(timeDiff) // don't forget the second param
-	var hours   = Math.floor(sec_num / 3600);
-	var minutes = Math.floor((sec_num - (hours * 3600)) / 60);
-	var seconds = sec_num - (hours * 3600) - (minutes * 60);
-	var uptimeHours = hours.toString(), uptimeMinutes = minutes.toString(), uptimeSeconds = seconds.toString();
+	timeDiff /= 1000;
 
-	if (hours   < 10) {uptimeHours   = "0"+hours;}
-	if (minutes < 10) {uptimeMinutes = "0"+minutes;}
-	if (seconds < 10) {uptimeSeconds = "0"+seconds;}
+	const sec_num = Math.round(timeDiff),
+		hours = Math.floor(sec_num / 3600),
+		minutes = Math.floor((sec_num - (hours * 3600)) / 60),
+		seconds = sec_num - (hours * 3600) - (minutes * 60);
 
-	let stats = JSON.parse(fs.readFileSync(path.join(__dirname, "../stats.json")));
+	let uptimeHours = hours.toString(), uptimeMinutes = minutes.toString(), uptimeSeconds = seconds.toString();
 
-	stats.uptime = uptimeHours+':'+uptimeMinutes+':'+uptimeSeconds;
+	if (hours < 10) uptimeHours = "0" + hours;
+	if (minutes < 10) uptimeMinutes = "0" + minutes;
+	if (seconds < 10) uptimeSeconds = "0" + seconds;
+
+	const stats = JSON.parse(fs.readFileSync(path.join(__dirname, "../stats.json")) as any);
+
+	stats.uptime = uptimeHours + ':' + uptimeMinutes + ':' + uptimeSeconds;
 	stats.cached_tracks = fs.readdirSync(path.join(__dirname, "../cache/data")).length
 
 	res.send({
@@ -62,23 +60,18 @@ server.get('/api/treble/stats', async (request, reply) => {
 	const { data } = await axios.get<Stats>(config.azura_api_url ?? "https://radio.bunker.dance/connect");
 
 	let fields = <any>data.now_playing.song.custom_fields;
-	let spotifyTrack:any;
-	if (fields.isrc) {
-		spotifyTrack = await lookup.search(`isrc:${fields.isrc}`)
-	} else {
-		spotifyTrack = await lookup.search(`track:${data.now_playing.song.title} artist:${data.now_playing.song.artist}`);
+	let spotifyTrack: any;
+	if (fields.isrc) spotifyTrack = await lookup.search(`isrc:${fields.isrc}`)
+	else spotifyTrack = await lookup.search(`track:${data.now_playing.song.title} artist:${data.now_playing.song.artist}`);
+
+	let track: any;
+	if (spotifyTrack) track = await lookup.getTrack(spotifyTrack);
+	else track = {
+		title: data.now_playing.song.title,
+		artist: data.now_playing.song.artist,
+		album: data.now_playing.song.album
 	}
 
-	let track:any;
-	if (spotifyTrack) {
-		track = await lookup.getTrack(spotifyTrack);
-	} else {
-		track = {
-			title: data.now_playing.song.title,
-			artist: data.now_playing.song.artist,
-			album: data.now_playing.song.album
-		}
-	}
 	reply.send({
 		success: true,
 		song: track,
@@ -100,43 +93,36 @@ server.get('/api/treble/stats', async (request, reply) => {
 
 server.get("/api/treble/lookup", async (req: any, res) => {
 	res.header("Access-Control-Allow-Origin", "*");
+
 	let search = req.query.search;
 
-	if (!search) {
-		res.send({
+	if (!search) return res.send({
+		success: false,
+		error: {
+			code: 404,
+			description: "Not Found"
+		}
+	})
+
+	const spotifyTrack = await lookup.search(search);
+
+	if (!spotifyTrack) {
+		search = search.replace(",", ", ");
+		search = search.replace("/", ", ");
+
+		const spotifyTrack = await lookup.search(search);
+
+		if (!spotifyTrack) return res.send({
 			success: false,
 			error: {
 				code: 404,
 				description: "Not Found"
 			}
 		})
-		return;
-	}
-	//get spotify track list
-	let spotifyTrack = await lookup.search(search);
-
-	//check if spotify is there
-	if (!spotifyTrack) {
-		//quality enhance
-		search = search.replace(",", ", ");
-		search = search.replace("/", ", ");
-
-		//try it again
-		spotifyTrack = await lookup.search(search);
-
-		if (!spotifyTrack) {
-			res.send({
-				success: false,
-				error: {
-					code: 404,
-					description: "Not Found"
-				}
-			})
-			return;
-		}
 	}
 
-	let track = await lookup.getTrack(spotifyTrack);
+	const track = await lookup.getTrack(spotifyTrack);
+
 	res.send({
 		success: true,
 		result: track
